chore: organise models
